requirements.txt
flwr==1.1.0
numpy==1.21.0
omegaconf==2.3.0
pandas==1.0.5
path.py==12.5.0
Pillow==9.3.0
pyrootutils==1.0.4
pytorch_lightning==1.8.6
ray==2.2.0
rich==12.6.0
torch==1.13.1
torchmetrics==0.11.0
torchvision==0.14.1
transformers==4.25.1
wandb==0.13.7
seaborn==0.12.2
hydra-core==1.3.2
hydra_colorlog==1.2.0
opencv-python==4.10.0.84
scikit-learn==1.0.2
pandarallel==1.6.5
wordcloud==1.9.4
stanza==1.6.1 
ipywidgets==8.1.5
umap-learn==0.5.7